+++ src/teal/sync.h
@@ -20,7 +20,7 @@ extern Sync sync;
 void sync_init(int *argc, char ***argv);
 void sync_reinit(MPI_Comm comm);
 
-void sync_exit(long status) __attribute((noreturn));
+void sync_exit(int status) __attribute((noreturn));
 
 long sync_lmin(long val);
 long sync_lmax(long val);
